requirements.txt
Flask
requests
mysql-connector
regex